+++ src/common/node15Internal.ts
@@ -1,0 +1,8 @@
+/*---------------------------------------------------------
+ * Copyright (C) Microsoft Corporation. All rights reserved.
+ *--------------------------------------------------------*/
+
+/**
+ * URL prefix that Node 15 and onwards uses for its internals.
+ */
+export const node15InternalsPrefix = 'node:';